+++ examples/example.py
@@ -1,0 +1,12 @@
+#
+# This file is part of Python Client Library for Sentinel-2 Angle Bands.
+# Copyright (C) 2021 INPE.
+#
+# Python Client Library for Sentinel-2 Angle Bands is free software; you can redistribute it and/or modify it
+# under the terms of the MIT License; see LICENSE file for more details.
+#
+"""Define a minimal example Sentinel-2 Angle Bands."""
+
+import s2angs
+
+s2angs.gen_s2_ang('/path/to/my/S2_file/S2B_MSIL1C_20191223T131239_N0208_R138_T23KMR_20191223T135458.zip')