+++ CHANGES.rst
@@ -1,6 +1,25 @@
-=======
+..
+    This file is part of Python Client Library for Sentinel-2 Angle Bands.
+    Copyright (C) 2021 INPE.
+
+    Python Client Library for Sentinel-2 Angle Bands is free software; you can redistribute it and/or modify it
+    under the terms of the MIT License; see LICENSE file for more details.
+
+
 Changes
 =======
 
-**0.0**
-Working Scripts
+
+Version 0.4.0 (2021-07-05)
+--------------------------
+
+- Restructure using cookiecutter
+- Support Angle Resampling for folder
+
+
+Released 2020
+
+
+- Source code versioning based on `Semantic Versioning 2.0.0 <https://semver.org/>`_.
+
+- License: `MIT <https://github.com/brazil-data-cube/sentinel2_angle_bands/blob/master/LICENSE>`_.